Run GC when saving snapshots
Snapshots in agent continued to grow, because only documents in clients
were subject to garbage collection.

This commit runs GC when saving snapshots to prevent it from continuing to grow.
